refactor(#48): corrige import de estilizações
